refactor: be more deliberate about exports
