+++ rust/benches/read_checkpoint.rs
@@ -1,5 +1,5 @@
 use criterion::{criterion_group, criterion_main, Criterion};
-use deltalake::state::DeltaTableState;
+use deltalake::table::state::DeltaTableState;
 use deltalake::DeltaTableConfig;
 use std::fs::File;
 use std::io::Read;
