+++ rust/src/delta_datafusion.rs
@@ -58,9 +58,9 @@ use object_store::ObjectMeta;
 use url::Url;
 
 use crate::action::{self, Add};
-use crate::builder::ensure_table_uri;
 use crate::errors::{DeltaResult, DeltaTableError};
 use crate::storage::ObjectStoreRef;
+use crate::table::builder::ensure_table_uri;
 use crate::table::state::DeltaTableState;
 use crate::{open_table, open_table_with_storage_options, DeltaTable, Invariant, SchemaDataType};
