+++ rust/src/operations/transaction/conflict_checker.rs
@@ -6,9 +6,9 @@ use object_store::ObjectStore;
 
 use super::CommitInfo;
 use crate::action::{Action, Add, DeltaOperation, MetaData, Protocol, Remove};
-use crate::config::IsolationLevel;
 use crate::errors::{DeltaResult, DeltaTableError};
 use crate::storage::commit_uri_from_version;
+use crate::table::config::IsolationLevel;
 use crate::table::state::DeltaTableState;
 
 #[cfg(feature = "datafusion")]
