+++ rust/src/storage/s3.rs
@@ -1,7 +1,7 @@
 //! AWS S3 storage backend.
 
 use super::utils::str_is_truthy;
-use crate::builder::{s3_storage_options, str_option};
+use crate::table::builder::{s3_storage_options, str_option};
 use bytes::Bytes;
 use dynamodb_lock::{DynamoError, LockClient, LockItem, DEFAULT_MAX_RETRY_ACQUIRE_LOCK_ATTEMPTS};
 use futures::stream::BoxStream;
